+++ FirebaseAppCheck/Tests/Unit/Core/FIRAppCheckIntegrationTests.m
@@ -18,7 +18,7 @@
 
 #import <OCMock/OCMock.h>
 
-@import FirebaseAppCheckInterop;
+#import <FirebaseAppCheckInterop/FirebaseAppCheckInterop.h>
 
 #import "FirebaseAppCheck/Sources/Core/TokenRefresh/FIRAppCheckTokenRefresher.h"
 #import "FirebaseAppCheck/Sources/Public/FirebaseAppCheck/FIRAppCheck.h"