+++ Firestore/Swift/Source/Codable/TimestampDecodingStrategy.swift
@@ -15,7 +15,11 @@
  */
 
 import Foundation
-import FirebaseFirestore
+#if SWIFT_PACKAGE
+  @_exported import FirebaseFirestoreInternalWrapper
+#else
+  @_exported import FirebaseFirestoreInternal
+#endif // SWIFT_PACKAGE
 import FirebaseSharedSwift
 
 public extension FirebaseDataDecoder.DateDecodingStrategy {