+++ ReleaseTooling/Sources/ZipBuilder/ModuleMapBuilder.swift
@@ -47,7 +47,7 @@ struct ModuleMapBuilder {
 
       """
 
-      if module == "FirebaseFirestore" {
+      if module == "FirebaseFirestoreInternal" {
         content += """
           link framework "BoringSSL-GRPC"
           link framework "gRPC-Core"