+++ scripts/check_firestore_symbols.sh
@@ -15,10 +15,11 @@
 # limitations under the License.
 
 # DESCRIPTION: This script identifies Objective-C symbols within the
-# `FirebaseFirestore.xcframework` that are not automatically linked when used
-# in a client target. Because the `FirebaseFirestore.xcframework` should
-# function without clients needing to pass the `-ObjC` flag, this script
-# catches potential regressions that break that requirement.
+# `FirebaseFirestoreInternal.xcframework` that are not automatically linked
+# when used in a client target. Because the
+# `FirebaseFirestoreInternal.xcframework` should function without clients
+# needing to pass the `-ObjC` flag, this script catches potential regressions
+# that break that requirement.
 #
 # DEPENDENCIES: This script depends on the given Firebase repo's `Package.swift`
 # using the `FIREBASECI_USE_LOCAL_FIRESTORE_ZIP` env var to swap the Firestore
@@ -57,8 +58,8 @@ fi
 # Check if the given xcframework path is valid.
 FIRESTORE_XCFRAMEWORK_PATH=$2
 
-if [ "$(basename $FIRESTORE_XCFRAMEWORK_PATH)" != 'FirebaseFirestore.xcframework' ]; then
-  echo "The given xcframework is not a FirebaseFirestore.xcframework."
+if [ "$(basename $FIRESTORE_XCFRAMEWORK_PATH)" != 'FirebaseFirestoreInternal.xcframework' ]; then
+  echo "The given xcframework is not a FirebaseFirestoreInternal.xcframework."
   exit 1
 fi