+++ FirebaseDatabaseSwift/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 10.17.0
+- [deprecated] All of the public API from `FirebaseDatabaseSwift` can now
+  be accessed through the `FirebaseDatabase` module. Therefore,
+  `FirebaseDatabaseSwift` has been deprecated, and will be removed in a
+  future release. See https://firebase.google.com/docs/ios/swift-migration for
+  migration instructions.
+
 # 9.0.0
 - [added] **Breaking change:** `FirebaseDatabaseSwift` has exited beta and is
   now generally available for use.