+++ FirebaseRemoteConfigSwift/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 10.17.0
+- [deprecated] All of the public API from `FirebaseRemoteConfigSwift` can now
+  be accessed through the `FirebaseRemoteConfig` module. Therefore,
+  `FirebaseRemoteConfigSwift` has been deprecated, and will be removed in a
+  future release. See https://firebase.google.com/docs/ios/swift-migration for
+  migration instructions.
+
 # 10.1.0
 - [fixed] Fix CocoaPods release did not include the RemoteConfigProperty feature. (#10371)